+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - This bug made it difficult to "stop" a channel after previewing it
 - Fix bug where deco default filler would never use hardware acceleration
 - Fix deleting local libraries with MySql backend
+- Fix `Scaling Behavior` `Crop` when content is smaller than FFmpeg Profile resolution
+  - Now, content will properly scale beyond the desired resolution before cropping
 
 ## [0.8.8-beta] - 2024-09-19
 ### Added